pyproject.toml
[tool.poetry]
name = "langchain-work01"
version = "0.1.0"
description = ""
authors = ["shouhei1310 <kdru1310@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
gradio = "4.44.1"
setuptools = "^77.0.3"
python-dotenv = "1.0.0"
langchain = "^0.3.21"
openai = "^1.68.0"
langchain-openai = "^0.3.9"
langchain-community = "^0.3.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
